// src/type/type.ts
export type Product ={
    id: number;
    title: string;
    description: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    brand: string;
    category: string;
    thumbnail: string;
    images: string[];
}
export type propCartItem= {
    product: Product;
    quantity: number;
  }